requirements.txt
# Original dependencies iteration_1
langchain-openai>=0.0.5
langchain-core>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0

# Additional dependencies for iteration_2
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic-ai>=0.0.1
logfire>=0.15.0
typing-extensions>=4.10.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-asgi>=0.42b0
python-multipart>=0.0.6
